Removed whole process/portal thing
It became obvious that a simple tap action on a button representing a deck type and assigning 30-50 tap actions for each deck (to move each card separately) was going to be necessary.  Even if I could have done mupltiple piece_types, this would still probably be the most straightforward and easiest to extend.